Run adaptive elastic net to identify CREs regulating target genes

This function will identify putative cis-regultory elements for each of the inputted genes. It takes as input your Seurat object, the pseudocell matrices generated by prepare_pseudocell_matrix, the list of genes that you wish to model, and parameters for running the adaptive elastic net and parallelization. Genes without genomic peaks (mitochondrial) in the provided Annotation are excluded, and names are printed if verbose is TRUE. The function returns information for each modeled gene as a list. Each gene's entry in this list is a list containing the following information: the coefficient of determination calculated during the cross validation step with the calculated lambda.min value, the trained model object, and a data frame with regulatory weights and standard errors for all statistically significant CREs. Regulatory weights for non-significant CREs are set to 0 for future analysis if desired.

Usage

run_peak_aen(
  seurat,
  pseudocell_mat,
  pseudocell_mat_peaks,
  gene_list,
  lambda2 = 0.5,
  gamma = 1,
  ci_cutoff = 1.96,
  pval_cutoff = NULL,
  set_nonsig_to_zero = TRUE,
  max_distance = 5e+05,
  train_fraction = 0.8,
  num_bootstraps = 1000,
  bootstrap = TRUE,
  num_threads = 16,
  globals_maxsize = NULL,
  verbose = TRUE,
  bs_seed = 943126,
  multi_seed = 6283408,
  ...
)

Arguments

seurat

A Seurat object

pseudocell_mat

Pseudocell expression matrix to use

pseudocell_mat_peaks

Pseudocell chromatin accessibility matrix to use

gene_list

Names of genes to model

lambda2

Parameter for quadratic L2 penalty for regularization

gamma

Positive constant for calculating adaptive weights

ci_cutoff

Confidence interval cutoff used to determine whether an estimated non-zero coefficient should be kept

pval_cutoff

As an alternative, p value cutoff to use in order to calculate ci_cutoff. E.g. p value of 0.5 corresponds to 95% CI.

set_nonsig_to_zero

Boolean to determine whether non-significant coefficient estimates should be set to zero

max_distance

Max distance from modeled gene's TSS to include CREs for building gene regulatory network model

train_fraction

Fraction of pseudocells used for training during cross-validation step

num_bootstraps

Number of bootstrap samples

bootstrap

Boolean if set to false, will not perform bootstrapping

num_threads

Number of threads to use for multithreading

globals_maxsize

If running into RAM problems during multithreading, can use this parameter to set future.globals.maxSize if not done manually

verbose

Boolean to determine whether to run silently

bs_seed

Seed for reproducibility in determining bootstrap seeds, set to NULL if not needed

multi_seed

Seed for reproducibility during multithreaded model training, set to NULL if not needed

Value

list of lists per modeled gene, containing modeling and CRE gene regulatory network information